Reorganized C++ library code into one implementation source file and one header file (instead of 3+3), changing some spacing and ordering, without changing functionality.
